health_pubs/pyproject.toml
[project]
name = "backend-alpha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.11.2",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyascii==0.3.2",
    "anyio==4.6.2.post1",
    "asgiref==3.8.1",
    "astroid==3.3.5",
    "asttokens==2.4.1",
    "attrs==24.2.0",
    "authlib==1.3.2",
    "autoflake==2.3.1",
    "awscli==1.35.20",
    "backcall==0.2.0",
    "bcp47==0.1.0",
    "bcrypt==4.2.0",
    "beautifulsoup4==4.12.3",
    "black==24.10.0",
    "bleach==6.2.0",
    "blis==1.0.1",
    "boto3==1.35.54",
    "botocore==1.35.54",
    "catalogue==2.0.10",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "cloudpathlib==0.20.0",
    "colorama==0.4.6",
    "confection==0.1.5",
    "coverage==7.6.4",
    "cryptography==43.0.3",
    "cymem==2.0.8",
    "decorator==4.4.2",
    "defusedxml==0.7.1",
    "dill==0.3.9",
    "distlib==0.3.9",
    "distro==1.9.0",
    "dj-rest-auth==7.0.0",
    "django-allauth==65.1.0",
    "django==5.1.2",
    "django-cors-headers==4.6.0",
    "django-extensions==3.2.3",
    "django-filter==24.3",
    "django-modelcluster==6.3",
    "django-permissionedforms==0.1",
    "django-taggit==6.1.0",
    "django-treebeard==4.7.1",
    "djangorestframework==3.15.2",
    "djangorestframework-jwt==1.11.0",
    "docopt==0.6.2",
    "docutils==0.16",
    "draftjs-exporter==5.0.0",
    "et-xmlfile==2.0.0",
    "exceptiongroup==1.2.2",
    "execnet==2.1.1",
    "executing==2.1.0",
    "fastjsonschema==2.20.0",
    "filelock==3.16.1",
    "filetype==1.2.0",
    "flake8==7.1.1",
    "frozenlist==1.5.0",
    "fsspec==2024.10.0",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httpx==0.27.2",
    "huggingface-hub==0.26.2",
    "identify==2.6.1",
    "idna==3.10",
    "imageio==2.36.0",
    "imageio-ffmpeg==0.5.1",
    "iniconfig==2.0.0",
    "ipython==8.12.3",
    "isort==5.13.2",
    "jedi==0.19.1",
    "jinja2==3.1.4",
    "jiter==0.7.0",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "jupyterlab-pygments==0.3.0",
    "l18n==2021.3",
    "laces==0.1.1",
    "langcodes==3.4.1",
    "language-data==1.2.0",
    "lxml==5.3.0",
    "marisa-trie==1.2.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mccabe==0.7.0",
    "mdurl==0.1.2",
    "mistune==3.0.2",
    "moviepy==1.0.3",
    "mpmath==1.3.0",
    "msal==1.31.0",
    "multidict==6.1.0",
    "murmurhash==1.0.10",
    "mutagen==1.47.0",
    "mypy-extensions==1.0.0",
    "nbclient==0.10.0",
    "nbconvert==7.16.4",
    "nbformat==5.10.4",
    "networkx==3.4.2",
    "nodeenv==1.9.1",
    "notifications-python-client==10.0.0",
    "numpy==2.1.3",
    "odfpy==1.4.1",
    "openai==0.28.0",
    "openpyxl==3.1.5",
    "packaging==24.1",
    "pandas==2.2.3",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pathlib-abc==0.1.1",
    "pathspec==0.12.1",
    "pathy==0.11.0",
    "pexpect==4.9.0",
    "pickleshare==0.7.5",
    "pillow==11.0.0",
    "pillow-heif==0.20.0",
    "pipdeptree==2.23.4",
    "pipreqs==0.5.0",
    "platformdirs==4.3.6",
    "pluggy==1.5.0",
    "podman-compose==1.2.0",
    "pre-commit==4.0.1",
    "preshed==3.0.9",
    "proglog==0.1.10",
    "prompt-toolkit==3.0.48",
    "propcache==0.2.0",
    "psycopg2-binary==2.9.10",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pycodestyle==2.12.1",
    "pycountry==24.6.1",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pyflakes==3.2.0",
    "pygments==2.18.0",
    "pyjwt>=1.7.1",
    "pylint==3.3.1",
    "pyodbc==5.2.0",
    "pypdf2==3.0.1",
    "pypng==0.20220715.0",
    "pytest==8.3.3",
    "pytest-cov==6.0.0",
    "pytest-django==4.9.0",
    "pytest-mock==3.14.0",
    "pytest-xdist==3.6.1",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "python-magic>=0.4.27",
    "python-pptx==1.0.2",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "pyzmq==26.2.0",
    "qrcode==8.0",
    "referencing==0.35.1",
    "regex==2024.9.11",
    "requests==2.32.3",
    "rich==13.9.4",
    "rpds-py==0.20.1",
    "rsa==4.7.2",
    "ruff==0.7.2",
    "s3transfer==0.10.3",
    "safetensors==0.4.5",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "segno==1.6.1",
    "sentencepiece==0.2.0",
    "shellingham==1.5.4",
    "six==1.16.0",
    "smart-open==6.4.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.36",
    "sqlparse==0.5.1",
    "srsly==2.4.8",
    "stack-data==0.6.3",
    "starlette==0.41.2",
    "sympy==1.13.3",
    "telepath==0.3.1",
    "thinc==9.1.1",
    "threadpoolctl==3.5.0",
    "tinycss2==1.4.0",
    "tokenizers==0.20.1",
    "tomli==2.0.2",
    "tomlkit==0.13.2",
    "tornado==6.4.1",
    "tqdm==4.66.6",
    "traitlets==5.14.3",
    "transformers==4.46.1",
    "typer==0.12.5",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.2.3",
    "uvicorn==0.32.0",
    "virtualenv==20.27.1",
    "wagtail==6.3",
    "wasabi==1.1.3",
    "wcwidth==0.2.13",
    "weasel==0.4.1",
    "webencodings==0.5.1",
    "willow==1.9.0",
    "wrapt==1.16.0",
    "xlsxwriter==3.2.0",
    "yarg==0.1.9",
    "yarl==1.17.1",
]

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | \.vscode
  | 0.*_initial\.py
)/
'''